+++ core/datastore/src/main/java/com/goms/datastore/serializer/AuthTokenSerializer.kt
@@ -1,7 +1,8 @@
-package com.goms.datastore
+package com.goms.datastore.serializer
 
 import androidx.datastore.core.CorruptionException
 import androidx.datastore.core.Serializer
+import com.goms.datastore.AuthToken
 import com.google.protobuf.InvalidProtocolBufferException
 import java.io.InputStream
 import java.io.OutputStream
