+++ core/network/src/main/java/com/goms/network/util/TokenAuthenticator.kt
@@ -1,6 +1,6 @@
 package com.goms.network.util
 
-import com.goms.datastore.AuthTokenDataSource
+import com.goms.datastore.datasource.auth.AuthTokenDataSource
 import com.goms.model.util.ResourceKeys
 import com.goms.network.BuildConfig
 import com.goms.network.api.AuthAPI
